+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for free-category
 
+## Version 0.0.2.0
+
+- EffCategory class and FreeEffCat category transformer
+- Example usage of FreeEffCat
+
 ## Version 0.0.1.0
 - free category (concrete and condensity transformed)
 - free arrows (concrete and condensity transformed)